+++ src/realm/group_shared.cpp
@@ -1451,9 +1451,18 @@ void SharedGroup::upgrade_file_format(bool allow_file_format_upgrade,
             if (!allow_file_format_upgrade)
                 throw FileFormatUpgradeRequired();
             gf::upgrade_file_format(m_group, target_file_format_version); // Throws
-            // Note: The file format version stored in in the Realm file will be
+            // Note: The file format version stored in the Realm file will be
             // updated to the new file format version as part of the following
             // commit operation. This happens in GroupWriter::commit().
+            if (m_upgrade_callback) {
+                try {
+                    m_upgrade_callback(current_file_format_version_2, target_file_format_version); // Throws
+                }
+                catch(...) {
+                    rollback();
+                    throw;
+                }
+            }
             commit(); // Throws
         }
         else {